remove legacy merit rule doctypes and consolidate merit generation logic into merit_service.py
