--- slcm/admission/dashboard_chart/merit_rule/merit_rule.json
@@ -1,33 +1,0 @@
-{
- "based_on": "creation",
- "chart_name": "merit rule",
- "chart_type": "Count",
- "creation": "2026-02-25 16:26:31.904997",
- "currency": "",
- "docstatus": 0,
- "doctype": "Dashboard Chart",
- "document_type": "Merit Rule",
- "dynamic_filters_json": "[]",
- "filters_json": "[[\"Merit Rule\",\"is_active\",\"=\",1]]",
- "group_by_type": "Count",
- "idx": 0,
- "is_public": 1,
- "is_standard": 0,
- "modified": "2026-02-25 16:26:31.904997",
- "modified_by": "Administrator",
- "module": "Admission",
- "name": "merit rule",
- "number_of_groups": 0,
- "owner": "Administrator",
- "parent_document_type": "",
- "roles": [],
- "show_values_over_chart": 1,
- "source": "",
- "time_interval": "Daily",
- "timeseries": 1,
- "timespan": "Last Year",
- "type": "Bar",
- "use_report_chart": 0,
- "value_based_on": "",
- "y_axis": []
-}
