--- slcm/admission/doctype/merit_component/merit_component.py
@@ -1,5 +1,0 @@
-import frappe
-from frappe.model.document import Document
-
-class MeritComponent(Document):
-	pass
