--- slcm/admission/doctype/merit_rule/test_merit_rule.py
@@ -1,22 +1,0 @@
-# Copyright (c) 2026, TFSS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests import IntegrationTestCase
-
-
-# On IntegrationTestCase, the doctype test records and all
-# link-field test record dependencies are recursively loaded
-# Use these module variables to add/remove to/from that list
-EXTRA_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
-IGNORE_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
-
-
-
-class IntegrationTestMeritRule(IntegrationTestCase):
-	"""
-	Integration tests for MeritRule.
-	Use this class for testing interactions between multiple components.
-	"""
-
-	pass
