--- slcm/admission/doctype/merit_rule_component/merit_rule_component.json
@@ -1,52 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2026-02-16 15:19:21.864140",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "component_type",
-  "weight",
-  "is_active"
- ],
- "fields": [
-  {
-   "fieldname": "component_type",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Component Type",
-   "options": "Merit Component",
-   "reqd": 1
-  },
-  {
-   "fieldname": "weight",
-   "fieldtype": "Float",
-   "in_list_view": 1,
-   "label": "Weight (%)",
-   "reqd": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "is_active",
-   "fieldtype": "Check",
-   "in_list_view": 1,
-   "label": "Is Active"
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2026-05-11 10:16:35.659472",
- "modified_by": "Administrator",
- "module": "Admission",
- "name": "Merit Rule Component",
- "owner": "Administrator",
- "permissions": [],
- "row_format": "Dynamic",
- "rows_threshold_for_grid_search": 20,
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
