--- slcm/admission/doctype/merit_rule_component/merit_rule_component.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2026, TFSS and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class MeritRuleComponent(Document):
-	pass
