--- slcm/admission/doctype/merit_rule_mapping/merit_rule_mapping.js
@@ -1,17 +1,0 @@
-
-
-frappe.ui.form.on("Merit Rule Mapping", {
-	refresh(frm) {
-
-
-	},
-	onload:function(frm){
-        frm.set_query("admission_cycle", function() {
-            return {
-                filters: {
-                    status: "Active"
-                }
-            };
-        });
-    },
-});
