--- slcm/admission/doctype/merit_rule_mapping/merit_rule_mapping.json
@@ -1,135 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2026-02-16 15:55:42.779170",
- "doctype": "DocType",
- "engine": "InnoDB",
- "field_order": [
-  "admission_cycle",
-  "campus",
-  "program_level",
-  "program",
-  "merit_rule",
-  "priority",
-  "is_active"
- ],
- "fields": [
-  {
-   "fieldname": "admission_cycle",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Admission cycle",
-   "options": "Admission Cycle",
-   "reqd": 1
-  },
-  {
-   "fieldname": "campus",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Campus",
-   "options": "Campus",
-   "reqd": 1
-  },
-  {
-   "fieldname": "program_level",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Programme level",
-   "options": "Undergraduate\nPostgraduate\nResearch Course",
-   "reqd": 1
-  },
-  {
-   "fieldname": "program",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Programme",
-   "options": "Program"
-  },
-  {
-   "fieldname": "merit_rule",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "Merit rule",
-   "options": "Merit Rule",
-   "reqd": 1
-  },
-  {
-   "fieldname": "priority",
-   "fieldtype": "Int",
-   "label": "Priority",
-   "reqd": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "is_active",
-   "fieldtype": "Check",
-   "label": "Is Active",
-   "reqd": 1
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "links": [],
- "modified": "2026-04-09 10:26:47.859566",
- "modified_by": "Administrator",
- "module": "Admission",
- "name": "Merit Rule Mapping",
- "naming_rule": "By script",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  },
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "Campus Admin",
-   "share": 1,
-   "write": 1
-  },
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "Merit Admin",
-   "share": 1,
-   "write": 1
-  },
-  {
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "Admission Admin",
-   "share": 1
-  }
- ],
- "row_format": "Dynamic",
- "rows_threshold_for_grid_search": 20,
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
