--- slcm/admission/doctype/merit_rule_mapping/test_merit_rule_mapping.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2026, TFSS and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestMeritRuleMapping(FrappeTestCase):
-	pass
