+++ slcm/admission/doctype/shortlisting_merit_list/shortlisting_merit_list.py
@@ -56,7 +56,7 @@ class ShortlistingMeritList(Document):
         self.total_shortlisted = len([a for a in self.shortlist_applicants if a.shortlist_status == "Shortlisted"])
         
         # Automatically execute shortlisting logic to fill categories
-        from slcm.admission.doctype.merit_rule.merit_service import execute_advanced_allocation_logic, _populate_category_lists
+        from slcm.admission.doctype.merit_generation.merit_service import execute_advanced_allocation_logic, _populate_category_lists
         execute_advanced_allocation_logic(self, is_shortlist_allocation=True)
         _populate_category_lists(self)
         
@@ -68,7 +68,7 @@ class ShortlistingMeritList(Document):
 
     @frappe.whitelist()
     def execute_shortlisting_logic(self):
-        from slcm.admission.doctype.merit_rule.merit_service import execute_advanced_allocation_logic, _populate_category_lists
+        from slcm.admission.doctype.merit_generation.merit_service import execute_advanced_allocation_logic, _populate_category_lists
         execute_advanced_allocation_logic(self, is_shortlist_allocation=True)
         _populate_category_lists(self)
         self.total_shortlisted = len([a for a in self.shortlist_applicants if a.shortlist_status == "Shortlisted"])
@@ -81,7 +81,7 @@ class ShortlistingMeritList(Document):
         """
         Triggers the Phase 2 Merit Generation (Entrance + Interview).
         """
-        from slcm.admission.doctype.merit_rule.merit_service import generate_merit_for_level
+        from slcm.admission.doctype.merit_generation.merit_service import generate_merit_for_level
         merit_list = generate_merit_for_level(
             self.admission_cycle, 
             self.campus,
