--- slcm/admission/number_card/merit_rule/merit_rule.json
@@ -1,26 +1,0 @@
-{
- "aggregate_function_based_on": "",
- "creation": "2026-02-20 16:10:31.844466",
- "currency": "",
- "docstatus": 0,
- "doctype": "Number Card",
- "document_type": "Merit Rule",
- "dynamic_filters_json": "[]",
- "filters_json": "[]",
- "function": "Count",
- "idx": 1,
- "is_public": 1,
- "is_standard": 1,
- "label": "Merit Rule",
- "modified": "2026-02-27 17:36:21.456158",
- "modified_by": "Administrator",
- "module": "Admission",
- "name": "Merit Rule",
- "owner": "Administrator",
- "parent_document_type": "",
- "report_function": "Sum",
- "show_full_number": 0,
- "show_percentage_stats": 1,
- "stats_time_interval": "Daily",
- "type": "Document Type"
-}
