--- slcm/admission/number_card/merit_rule_mapping/merit_rule_mapping.json
@@ -1,26 +1,0 @@
-{
- "aggregate_function_based_on": "",
- "creation": "2026-02-21 10:44:33.582228",
- "currency": "",
- "docstatus": 0,
- "doctype": "Number Card",
- "document_type": "Merit Rule Mapping",
- "dynamic_filters_json": "[]",
- "filters_json": "[]",
- "function": "Count",
- "idx": 0,
- "is_public": 1,
- "is_standard": 1,
- "label": "Merit Rule Mapping",
- "modified": "2026-02-27 17:36:31.211922",
- "modified_by": "Administrator",
- "module": "Admission",
- "name": "Merit Rule Mapping",
- "owner": "Administrator",
- "parent_document_type": "",
- "report_function": "Sum",
- "show_full_number": 0,
- "show_percentage_stats": 1,
- "stats_time_interval": "Daily",
- "type": "Document Type"
-}
