+++ slcm/workspace_sidebar/settings.json
@@ -185,28 +185,7 @@
    "show_arrow": 0,
    "type": "Section Break"
   },
-  {
-   "child": 1,
-   "collapsible": 1,
-   "indent": 0,
-   "keep_closed": 0,
-   "label": "Merit Rule",
-   "link_to": "Merit Rule",
-   "link_type": "DocType",
-   "show_arrow": 0,
-   "type": "Link"
-  },
-  {
-   "child": 1,
-   "collapsible": 1,
-   "indent": 0,
-   "keep_closed": 0,
-   "label": "Merit Rule Mapping",
-   "link_to": "Merit Rule Mapping",
-   "link_type": "DocType",
-   "show_arrow": 0,
-   "type": "Link"
-  },
+
   {
    "child": 1,
    "collapsible": 1,
